Refactor: Centralize get_calendar_events function
Moved the get_calendar_events function to a new file `calendar-functions.php`
and updated `download-flyer.php`, `generate-post-image.php`, and
`image-generator-ui.php` to use this shared function.

This removes code duplication and improves maintainability.
